+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changes:
 
 - Added `childstate_getKeysPaged` RPC
 - Align static metadata with latest Substrate
+- Adjust error message for contracts V3 (extrinsic from runtime)
 
 
 ## 4.14.1 Jun 14, 2021