+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Changes:
 
 - Added `childstate_getKeysPaged` RPC
+- Add `XcmOrigin` mapping for `OriginCaller` generation
 - Align static metadata with latest Substrate
 - Adjust error message for contracts V3 (extrinsic from runtime)