+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## master
+## 4.15.1 Jun 19, 2021
 
 Upgrade priority: Medium. Recommended for all chains where Xcm is used (addition of new `XcmOrigin`)